Move sklearn tests to a separate file
